ci: Add debug option to kiwi build action
This commit introduces a new input parameter, 'kiwi-debug', to the
kiwi-build GitHub Action. This parameter allows users to enable or
disable debug mode when building a Kiwi image.

In the `action.yml` file, the 'kiwi-debug' input is added, with a
required status and a default empty string. The `build-kiwi-container.sh`
script updates to check if 'kiwi-debug' is specified. If it is, the
`--debug` flag is appended to the `kiwi-ng` command to enable debug mode.

This enhancement provides more control over the build process,
facilitating easier troubleshooting and debugging of Kiwi builds.