ci: Fix typo in action.yml input parameter
Correct a typo in the "kiwi-extra-args" input parameter name in the
"kiwi-build/action.yml" file. The label had an extra colon at the end,
preventing the correct parsing of this input by the build system.

The change ensures the action can be executed correctly by removing
the extraneous colon that was causing the issue:

```yaml
- kiwi-extra-args::
+ kiwi-extra-args:
```